adding random logic for empty fields
